Move superseded files to _deprecated, require apex/quack directly
- Move dali_imagenet.py, dali_imagenet_optimized.py to
  experiments/datamodules/_deprecated/
- Move vit5_small_pretrain_apex_dali.py,
  vit5_small_pretrain_apex_dali_optimized_plus.py to
  examples/vit5_imagenet/_deprecated/
- Remove try/except ImportError fallbacks for apex FusedLAMB in all
  active configs — import directly, fail loudly if not installed
- Remove try/except for quack rmsnorm — import directly
- Update profile_training_bottleneck.py imports to use _deprecated paths

Required packages (installed in nvsubquadratic_cuda129.sqsh):
  apex 0.1, quack-kernels 0.2.10, nvidia-dali-cuda120 1.53.0

Made-with: Cursor
