+++ examples/vit5_imagenet/_deprecated/vit5_small_pretrain_apex_dali.py
