+++ experiments/datamodules/_deprecated/dali_imagenet.py
